Hillel

+++ src/main/java/com/example/springhillel/repository/RoleRepository.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface RoleRepository {
 
     void createRole(Role role);
-    void roleAssignment(int userId, int roleId);
+    void roleAssignment(long userId, long roleId);
     void addActionPontToRole(int roleId, int actionId);
-    void deleteRole(int id);
+    void deleteRole(long id);
     List<Role> getActionPointRole();
 }
