+++ src/main/java/com/example/springhillel/repository/jdbcRepository/JdbcRoleRepositoryImpl.java
@@ -22,7 +22,7 @@ public class JdbcRoleRepositoryImpl implements RoleRepository {
     }
 
     @Override
-    public void roleAssignment(int userId, int roleId) {
+    public void roleAssignment(long userId, long roleId) {
         jdbcTemplate.update("UPDATE user set role_id = ? where id = ?",
                 userId, roleId);
     }
@@ -34,7 +34,7 @@ public class JdbcRoleRepositoryImpl implements RoleRepository {
     }
 
     @Override
-    public void deleteRole(int id) {
+    public void deleteRole(long id) {
         jdbcTemplate.update("DELETE FROM user_role WHERE id = ?", id);
 
     }
